edge/pi/services/old/mva/requirements.txt
paho-mqtt==2.1.0
numpy==1.26.4
sqlite-utils==3.36
python-dateutil==2.9.0
python-dotenv==1.0.1
onnxruntime==1.18.0